HIve_query-1
This HQL file shows the process of uploading the raw data on a Hive Warehouse. It uploads all the data in the train.csv file to the Hive Warehouse from this we can use this raw data to extract the desired rows and columns and perform Data Cleaning by removing the Null values. After completion of this we can see the NULL value in the table it is because Hive substitutes the value to null if it does not follow the same as the Datatype mentioned during the creation of the table i.e. if the Value of Age was assigned the Datatype Integer and if some value in the datatype of Float it will not accept that and pull in place of that.