Hive_query-2
Here when you implement this HQL file it will remove the NULL values in the age section and replace it with the average value of the whole dataset of age that is 29(Average of the whole data set is 29.84 and after this replacement, it comes out to be 29.64). All this transformation were implemented on a View instead of table and they were stored in a final table which can be used for the generation of the Analysis in Tableau.